Allen/contract struct

+++ contracts/tests/ExposedMultiToken.sol
@@ -5,7 +5,7 @@ import "../base/MultiToken.sol";
 
 contract ExposedMultiToken is MultiToken {
     constructor(string memory _postfix) MultiToken(_postfix) {}
-    function createToken(address _contractAddress, string memory _name, uint256 _decimal) public returns (uint) {
+    function createToken(address _contractAddress, string memory _name, uint8 _decimal) public returns (uint) {
         return _createToken(_contractAddress, _name, _decimal);
     }
 }
